Load BGE encoder in FP16 to halve VRAM usage and speed up encoding

+++ src/embedding_model.py
@@ -40,7 +40,9 @@ class EmbeddingModel:
 
         else:  # wikipedia
             self._query_tokenizer = AutoTokenizer.from_pretrained(BGE_ENCODER)
-            self._query_model = AutoModel.from_pretrained(BGE_ENCODER).to(device).eval()
+            self._query_model = AutoModel.from_pretrained(
+                BGE_ENCODER, torch_dtype=torch.float16
+            ).to(device).eval()
 
     # ------------------------------------------------------------------
     # Public API
